Implementation of SASS
