+++ package.json
@@ -15,7 +15,8 @@
     "@mui/material-nextjs": "^5.15.11",
     "next": "14.2.3",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "sass": "^1.77.5"
   },
   "devDependencies": {
     "@types/node": "^20",
