+++ src/app/auth/components/auth-background/auth-background.scss
@@ -1,0 +1,9 @@
+.auth-background {
+    position: absolute;
+    z-index: -1;
+    bottom: 0;
+    left: 0;
+    top: 0;
+    right: 0;
+    overflow: hidden
+}
