+++ src/app/auth/components/auth-background/auth-background.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import Box from '@mui/material/Box';
+
+const AuthBackground = () => {
+
+    return (
+        <Box className='auth-background' />
+    );
+};
+
+export default AuthBackground;
