+++ src/app/auth/components/auth-card/auth-card.scss
@@ -1,0 +1,18 @@
+.auth-card {
+    padding: 8px;
+
+    /* Small devices */
+    @media (min-width: 600px) {
+        padding: 12px;
+    }
+
+    /* Medium devices */
+    @media (min-width: 960px) {
+        padding: 16px;
+    }
+
+    /* Large devices */
+    @media (min-width: 1200px) {
+        padding: 20px;
+    }
+}
