+++ src/app/auth/components/auth-wrapper/auth-wrapper.scss
@@ -1,0 +1,27 @@
+.auth-wrapper {
+
+    min-height: 100vh;
+
+    .grid1 {
+        flex-direction: column;
+        justify-content: center;
+        min-height: 100vh;
+    }
+
+    .grid2 {
+        justify-content: center;
+        align-items: center;
+
+        @media (min-width: 1024px) {
+            min-height: calc(100vh - 112px);
+        }
+
+        @media (min-width: 768px) {
+            min-height: calc(100vh - 134px);
+        }
+
+        @media (min-width: 0px) {
+            min-height: calc(100vh - 210px);
+        }
+    }
+}
