+++ src/app/auth/sign-in/page.tsx
@@ -8,14 +8,13 @@ import Typography from "@mui/material/Typography";
 const SignIn = () => {
   return (
     <Grid container spacing={3}>
-
       <Grid item xs={12}>
         <Stack
           direction="row"
-          justifyContent="space-between"
+          justifyContent="center"
           alignItems="baseline"
           sx={{ mb: { xs: -0.5, sm: 0.5 } }}>
-          <Typography variant="h3">Sign In</Typography>
+          <Typography variant="h5">Login</Typography>
         </Stack>
       </Grid>
